+++ packages/core/src/lib/PopoverDropdown/PopoverDropdown.tsx
@@ -2,6 +2,7 @@ import { IPopoverDropdownProps } from './types';
 import styles from './PopoverDropdown.styles';
 import { color } from '@web3uikit/styles';
 import depreciatedWarning from '../../utils/depreciationWarning';
+import { Fragment } from 'react';
 
 const { DivStyled, ListStyled } = styles;
 
@@ -27,9 +28,9 @@ const PopoverDropdown: React.FC<IPopoverDropdownProps> = ({
             id={id}
             {...props}
         >
-            {parent}
-
+            <Fragment key="popover-dropdown-parent">{parent}</Fragment>
             <ListStyled
+                key="popover-dropdown-children"
                 backgroundColor={backgroundColor}
                 data-testid="test-popover-dropdown-list"
                 moveBody={moveBody}